Add CI, fix a mut_from_ref unsoundness smell, clear clippy
CI runs three jobs: the weightless suite on ubuntu x86-64 and macOS arm64 so
both dispatch paths are exercised, fmt plus clippy -D warnings, and a
bit-identity job that clones the C reference at the pinned commit, builds it,
runs both tiny-model oracles and byte-compares the final-position logits. The
exactness claim is now machine-checked on every push instead of asserted once.

cargo clippy was FAILING, not merely warning: clippy::mut_from_ref is
deny-by-default and SlotWriter::slot took &self and returned &mut [u8], which
manufactures a mutable borrow from a shared one. The aliasing was legitimate -
phase 1 reserves each slot exclusively - but the signature let any two holders
of the same &SlotWriter obtain &mut to the same bytes with nothing objecting.
It now hands out a raw pointer and the slice is built at the call site, where
the reservation argument already lives. My earlier check only counted lines
starting with 'warning:', so the error slipped past.

The remaining 100 clippy warnings are cleared: div_ceil, Error::other,
saturating_sub, sort_by_key, and the rest of the mechanical simplifications.
needless_range_loop and too_many_arguments are allowed crate-wide with a
rationale, because the index loops and the signatures are what keep this
diffable against the C source.

Verified after: clippy -D warnings exits 0, fmt clean, 43 tests pass, and the
logits are still byte-identical to the C build.

Also corrected NOTICE. It was upstream's file verbatim, naming third_party/
paths that do not exist here and a tech-report PDF that was never copied. It
now states the derivative-work relationship, the pinned upstream commit, what
was taken and what is original, and reproduces the upstream NOTICE in full as
section 4(d) requires.

+++ src/bind.rs
@@ -300,7 +300,7 @@ impl LayerPlan {
                     }
                     let dst = &mut widen[w..w + need];
                     let rp = &run[s.off as usize..];
-                    let rowb = (4 + cols as usize) as usize;
+                    let rowb = 4 + cols as usize;
                     for r in 0..rows as usize {
                         let scale = f32::from_le_bytes([
                             rp[r * rowb],
